double client works fine

+++ srcs/requirements/backend/src/sockets/sockets-friends.gateway.ts
@@ -29,53 +29,65 @@ export class SocketsFriendsGateway implements OnGatewayConnection, OnGatewayDisc
 
   @SubscribeMessage('getFriend')
   async getFriendsByUserId(@ConnectedSocket() client: Socket) {
-    const user = this.socketService.getUser(client.id);
+    const user = await this.socketService.getUser(client.id);
+    if (!user)
+      return ;
     const friends = await this.friendService.getFriendsByUserId(user.id);
     return friends;
   }
 
   @SubscribeMessage('getFriendReq')
   async getFriendReq(@ConnectedSocket() client: Socket) {
-    const user = this.socketService.getUser(client.id);
+    const user = await this.socketService.getUser(client.id);
+    if (!user)
+      return ;
     return this.friendService.getFriendReq(user.id)
   }
 
   @SubscribeMessage('send')
   async sendFriendReq(@ConnectedSocket() client: Socket, @MessageBody() body: { id: number }) {
-    const user = this.socketService.getUser(client.id);
+    const user = await this.socketService.getUser(client.id);
+    if (!user)
+      return ;
     const request = await this.friendService.createFriendRequest(user.id, +body.id)
     if (!request)
       return;
-    const friendSocketId = this.socketService.findSocketById(+body.id)
-    if (friendSocketId) {
-      const friendSocket = this.server.sockets.sockets.get(friendSocketId)
-      if (friendSocket) {
-        friendSocket.emit('friendRequestNotification', { req: request });
-      }
-    }
+    // const friendSocketId = this.socketService.findSocketById(+body.id)
+    // if (friendSocketId) {
+    //   const friendSocket = this.server.sockets.sockets.get(friendSocketId)
+    //   if (friendSocket) {
+    //     friendSocket.emit('friendRequestNotification', { req: request });
+    //   }
+    // }
+    this.server.to(`user_${body.id}`).emit('friendRequestNotification', { req: request });
   }
 
   @SubscribeMessage('acceptFriend')
   async acceptFriendReq(@ConnectedSocket() client: Socket, @MessageBody() body: { id: number }) {
-    const user = this.socketService.getUser(client.id);
+    const user = await this.socketService.getUser(client.id);
+    if (!user)
+      return ;
     await this.friendService.acceptFriendRequest(+user.id, +body.id)
-    const friendSocketId = this.socketService.findSocketById(+body.id)
+    // const friendSocketId = this.socketService.findSocketById(+body.id)
     const friendFriend = await this.friendService.getFriendsByUserId(+body.id)
     const friendUser = await this.friendService.getFriendsByUserId(+user.id)
     const request = this.friendService.getFriendReq(user.id)
     client.emit('receiveFriend', { friends: friendUser })
     client.emit('receiveReq', { req: request })
-    if (friendSocketId) {
-      const friendSocket = this.server.sockets.sockets.get(friendSocketId)
-      if (friendSocket) {
-        friendSocket.emit('receiveFriend', { friends: friendFriend });
-      }
-    }
+    // if (friendSocketId) {
+    //   const friendSocket = this.server.sockets.sockets.get(friendSocketId)
+    //   if (friendSocket) {
+    //     friendSocket.emit('receiveFriend', { friends: friendFriend });
+    //   }
+    // }
+      this.server.to(`user_${body.id}`).emit('receiveFriend', { friends: friendFriend });
   }
 
   @SubscribeMessage('refuseFriend')
   async refuseFriendReq(@ConnectedSocket() client: Socket, @MessageBody() body: { id: number }) {
-    const user = this.socketService.getUser(client.id);
+    const user = await this.socketService.getUser(client.id);
+    if (!user)
+      return ;
     await this.friendService.refuseFriendRequest(+user.id, +body.id)
     const request = this.friendService.getFriendReq(user.id)
     client.emit('receiveReq', { req: request })
@@ -83,23 +95,28 @@ export class SocketsFriendsGateway implements OnGatewayConnection, OnGatewayDisc
 
   @SubscribeMessage('deleteFriend')
   async deleteFriend(@ConnectedSocket() client: Socket, @MessageBody() body: { id: number }) {
-    const user = this.socketService.getUser(client.id);
+    const user = await this.socketService.getUser(client.id);
+    if (!user)
+      return ;
     await this.friendService.deleteFriendById(+user.id, +body.id)
     const friendUser = await this.friendService.getFriendsByUserId(+user.id)
     client.emit('receiveFriend', { friends: friendUser })
     const friendFriend = await this.friendService.getFriendsByUserId(+body.id)
-    const friendSocketId = this.socketService.findSocketById(+body.id)
-    if (friendSocketId) {
-      const friendSocket = this.server.sockets.sockets.get(friendSocketId)
-      if (friendSocket) {
-        friendSocket.emit('receiveFriend', { friends: friendFriend });
-      }
-    }
+    // const friendSocketId = this.socketService.findSocketById(+body.id)
+    // if (friendSocketId) {
+    //   const friendSocket = this.server.sockets.sockets.get(friendSocketId)
+    //   if (friendSocket) {
+    //     friendSocket.emit('receiveFriend', { friends: friendFriend });
+    //   }
+    // }
+    this.server.to(`user_${body.id}`).emit('receiveFriend', { friends: friendFriend });
   }
 
   @SubscribeMessage('bloqueUser')
   async bloqueUser(@ConnectedSocket() client: Socket, @MessageBody() body: { id: number }) {
-    const user = this.socketService.getUser(client.id);
+    const user = await this.socketService.getUser(client.id);
+    if (!user)
+      return ;
     await this.friendService.bloqueUserById(+user.id, +body.id)
     await this.friendService.refuseFriendRequest(+user.id, +body.id)
     const friendUser = await this.friendService.getFriendsByUserId(+user.id)
@@ -110,52 +127,64 @@ export class SocketsFriendsGateway implements OnGatewayConnection, OnGatewayDisc
 
   @SubscribeMessage('bloqueFriend')
   async bloqueFriend(@ConnectedSocket() client: Socket, @MessageBody() body: { id: number }) {
-    const user = this.socketService.getUser(client.id);
+    const user = await this.socketService.getUser(client.id);
+    if (!user)
+      return ;
     await this.friendService.bloqueUserById(+user.id, +body.id)
     const friendUser = await this.friendService.getFriendsByUserId(+user.id)
     client.emit('receiveFriend', { friends: friendUser })
     const friendFriend = await this.friendService.getFriendsByUserId(+body.id)
-    const friendSocketId = this.socketService.findSocketById(+body.id)
-    if (friendSocketId) {
-      const friendSocket = this.server.sockets.sockets.get(friendSocketId)
-      if (friendSocket) {
-        friendSocket.emit('receiveFriend', { friends: friendFriend });
-      }
-    }
+    // const friendSocketId = this.socketService.findSocketById(+body.id)
+    // if (friendSocketId) {
+    //   const friendSocket = this.server.sockets.sockets.get(friendSocketId)
+    //   if (friendSocket) {
+    //     friendSocket.emit('receiveFriend', { friends: friendFriend });
+    //   }
+    // }
+    this.server.to(`user_${body.id}`).emit('receiveFriend', { friends: friendFriend });
   }
 
   @SubscribeMessage('invitePlay')
   async invitePlay(@ConnectedSocket() client: Socket, @MessageBody() body: { id: number }) {
-    const user = this.socketService.getUser(client.id);
-    const friendSocketId = this.socketService.findSocketById(+body.id)
-    if (friendSocketId) {
-      const friendSocket = this.server.sockets.sockets.get(friendSocketId)
-      if (friendSocket) {
-        friendSocket.emit('invitePlayReq', { friendId: user.id });
-      }
-    }
+    const user = await this.socketService.getUser(client.id);
+    if (!user)
+      return ;
+    // const friendSocketId = this.socketService.findSocketById(+body.id)
+    // if (friendSocketId) {
+    //   const friendSocket = this.server.sockets.sockets.get(friendSocketId)
+    //   if (friendSocket) {
+    //     friendSocket.emit('invitePlayReq', { friendId: user.id });
+    //   }
+    // }
+    this.server.to(`user_${body.id}`).emit('invitePlayReq', { friendId: user.id });
+
   }
 
   @SubscribeMessage('acceptInvitation')
   async acceptInvit(@ConnectedSocket() client: Socket, @MessageBody() body: { id: number }) {
-    const user = this.socketService.getUser(client.id);
-    const friendSocketId = this.socketService.findSocketById(+body.id)
-    if (friendSocketId) {
-      const friendSocket = this.server.sockets.sockets.get(friendSocketId)
-      if (friendSocket) {
+    const user = await this.socketService.getUser(client.id);
+    if (!user)
+      return ;
+    // const friendSocketId = this.socketService.findSocketById(+body.id)
+    // if (friendSocketId) {
+    //   const friendSocket = this.server.sockets.sockets.get(friendSocketId)
+    //   if (friendSocket) {
+    //     const friend = await this.queueService.retUser(user.id)
+    //     friendSocket.emit('accepted', { friend });
+    //   }
+    // }
         const friend = await this.queueService.retUser(user.id)
-        friendSocket.emit('accepted', { friend });
+        this.server.to(`user_${body.id}`).emit('accepted', { friend });
       }
-    }
-  }
   @SubscribeMessage('refuseInvitation')
   async refuseInvit(@ConnectedSocket() client: Socket, @MessageBody() body: { id: number }) {
-    const friendSocketId = this.socketService.findSocketById(+body.id)
-    if (friendSocketId) {
-      const friendSocket = this.server.sockets.sockets.get(friendSocketId)
-      if (friendSocket) {
-        friendSocket.emit('refused');
-      }
-    }
+    // const friendSocketId = this.socketService.findSocketById(+body.id)
+    // if (friendSocketId) {
+    //   const friendSocket = this.server.sockets.sockets.get(friendSocketId)
+    //   if (friendSocket) {
+    //     friendSocket.emit('refused');
+    //   }
+    // }
+     this.server.to(`user_${body.id}`).emit('refused');
   }
 }
